Создал Product и кое что подправил из прошлого

+++ app/Models/Category.php
@@ -13,7 +13,7 @@ class Category extends Model
          'filters' => 'array',
          'tags_ru' => 'array',
          'tags_uk' => 'array',
-         'skidka' => 'integer',
+         'skidka' => 'float',
      ];
      protected $guarded = [];
